Calibration files can now be run-depenent
If the calibration object or file contains a key "calibrationList", it will loop thorough all entries in the array and look for the "best" match. Exact run numbers ("run":runNumber) are higher priority than ranges of run numbers ("runRange":[minRun maxRun] inclusive) which are higher prioirty then the default (no "run" or "runRange" keys). It will take the first match in the highest priorty.